+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - The following breaking changes only affect you if you're creating your own `BigNumber` subclasses:
   - the return type of `BigNumber::of()` is now `static`
   - `BigNumber` has a new abstract method `from()`
+  - all `public` and `protected` functions of `BigNumber` are now `final`
 
 ## [0.11.0](https://github.com/brick/math/releases/tag/0.11.0) - 2023-01-16